Update Changes; make all changed $VERSION values consistent
If we make changes to the module we need to update its $VERSION (obviously).
But if we change the $VERSION, change it to the version of the distribution
(which is $Cron::Sequencer::VERSION, which needs to be bumped for all
releases anyway).

Other routes lead to madness. Or at least "confusion and delay".