Fix --hide-env in the default (text) output
This is a regression introduced by commit 292f296d7fd91798:
    Add a --no-group option, that disables the default event grouping

Regression testing the output remains a TODO, and this bug shows that the
current interim solution of local "golden results" is fragile.

Not addressing testing *yet* - prioritising getting this fix shipped, rather
than delaying shipping the fix until I have time to decide on how and how much
to test the output.